[modules/spaceapi] Fix ArgumentException
This ArgumentException was caused by me failing to rename one occurence
of a parameter when refactoring code. This went under my radar as the
API I'm usually testing against, was offline at that time, too. So I
expceted to see an Error. Just not this one.

Well, the fix comes late, but better late than never :)

Signed-off-by: Tobias Manske <tobias.manske@mailbox.org>