Add support for nowplaying-tf, which provides better format strings
This is added with the tf_format config, which will override format if
it is set. This makes the call "deadbeef --nowplaying-tf <tf_format>"
and displays the output directly. This output should now also support
unicode, regardless of Python version.

There is also the tf_format_if_stopped, a boolean config that
configures whether or not the tf_format output should be displayed
even if deadbeef isn't playing anything.