add a parameter to control bluetooth popup menu
The right-click popup menu in the bluetooth module does not work well in
Sway (a i3wm-compatible Wayland WM). This is known issue to
Sway. Given that the popup menu is simply to turn on/off the
bluetooth device, the graphical popup is really not necessary. This
commit adds a parameter to control whether a popup menu should be used
when a right-click happens, or we simply turn on/off the device without
the menu.