[bumblebee] add small sleep before update
in some situations/modules (spotify, playerctl, etc.), it is possible
that a state change (e.g. play/pause) takes a small time to actually
propagate into the whole system (e.g. until the next "update" retrieves
the correct value).

To alleviate that, add a very small delay to the update.